events now have to arguments of size bit<32> and bit<24> to make use of the total 56 bits

+++ src/hardware_dep/dpdk/includes/dpdk_model_v1model.h
@@ -181,5 +181,5 @@ void timer_burst(uint32_t id, SHORT_STDPARAMS);
 //                     packet_descriptor_t* pd, lookup_table_t** tables);
 
 #ifdef EVENT_MODULE
-void raise_event(uint8_t *event_id, uint32_t *arg, SHORT_STDPARAMS);
+void raise_event(uint8_t *event_id, uint64_t *arg, SHORT_STDPARAMS);
 #endif
